% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_quarto}
\alias{format_quarto}
\alias{format.quarto_section}
\alias{format.quarto_tabset}
\alias{format.quarto_div}
\alias{format.quarto_span}
\alias{format.quarto_markdown}
\alias{format.quarto_group}
\title{Format a quarto object}
\usage{
\method{format}{quarto_section}(x, ...)

\method{format}{quarto_tabset}(x, ...)

\method{format}{quarto_div}(x, ...)

\method{format}{quarto_span}(x, ...)

\method{format}{quarto_markdown}(x, ...)

\method{format}{quarto_group}(x, ...)
}
\arguments{
\item{x}{The quarto object}

\item{...}{Other arguments (ignored)}
}
\value{
A formatted quarto object, sometimes a single string
(e.g., for \code{quarto_section} objects), but can be a list of
strings and/or plot objects (e.g., for \code{quarto_tabset} objects)
}
\description{
Creates a formatted representation of a quarto object in a
form suitable for printing.
}
\details{
The intent behind the \code{format()} methods for quarto objects
is to create a ready-to-print representation of that is almost
identical to what will be printed into the quarto document
when the print method is called. Because of this, the formatted
version of a quarto object is usually a string or a list of
strings, but it can also include plot objects that have not
yet been rendered. The resulting representation isn't usually
very pretty, though if passed to \code{cat()} it is generally
readable.
}
\examples{
# formatted sections, spans and divs
sec <- quarto_section("Header", .level = 2L)
spn <- quarto_span("Content", .class = "underline")
div <- quarto_div("Content", .class = "content-margin")

format(sec)
format(spn)
format(div)

# formatted tabsets
tbs <- quarto_tabset(
  .content = list(tab1 = 1:10, tab2 = "hello"),
  .title = "Header",
  .level = 2L
)

format(tbs)

# formatted groups and markdown

mkd <- quarto_markdown(list("- this is a", "- markdown list"), .sep = "\n")
gps <- quarto_group(list(div, mkd))

format(mkd)
format(gps)

}
