% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{groups}
\alias{groups}
\alias{quarto_output}
\alias{quarto_markdown}
\alias{quarto_paragraph}
\title{Quarto groups}
\usage{
quarto_output(...)

quarto_markdown(..., .sep = "")

quarto_paragraph(..., .sep = "")
}
\arguments{
\item{..., }{Content}

\item{.sep, }{Separator}
}
\description{
Quarto groups
}
