% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\title{Insert section header}
\usage{
header(.title, .level)
}
\arguments{
\item{.title}{Header title}

\item{.level}{Numeric specfying the header level}
}
\description{
Places a section header in a quarto document
}
