% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.quarto_object}
\alias{knit_print.quarto_object}
\title{Print a quarto object with knitr}
\usage{
\method{knit_print}{quarto_object}(x, ...)
}
\arguments{
\item{x}{A quarto object}

\item{...}{Other arguments (ignored)}
}
\value{
Invisibly returns NULL.
}
\description{
Prints a quarto object. When calling \code{knitr::knit_print()}
on a quarto object, the relevant \code{format()} method is called
first, and the formatted version is printed to the document.
}
\details{
Additional details...
}
