% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\alias{open_output}
\alias{close_output}
\title{Insert output sections}
\usage{
open_output()

close_output()
}
\description{
Opens and closes an output section in a quarto document
}
