% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_div}
\alias{quarto_div}
\title{Quarto divs}
\usage{
quarto_div(..., .class, .sep = "")
}
\arguments{
\item{..., }{Content to be contained in the div}

\item{.class, }{CSS class (or vector of classes) for the div}

\item{.sep, }{Separator}
}
\description{
Quarto divs
}
