% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_section}
\alias{quarto_section}
\title{Quarto section header}
\usage{
quarto_section(title, level)
}
\arguments{
\item{title, }{Section title}

\item{level, }{Header level}
}
\description{
Quarto section header
}
\examples{
quarto_section("A level-two header", level = 2L)
quarto_section("A level-three header", level = 3L)

}
