% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_span}
\alias{quarto_span}
\title{Quarto span}
\usage{
quarto_span(.content, .class = NULL)
}
\arguments{
\item{.content, }{Content to be contained in the span}

\item{.class, }{CSS class (or vector of classes) for the span}
}
\description{
Quarto span
}
