% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_tabsec}
\alias{quarto_tabsec}
\title{Quarto tabset with section header}
\usage{
quarto_tabsec(.content, .title, .level = 3L)
}
\arguments{
\item{.content, }{Tab title and content as named list}

\item{.title, }{Section title}

\item{.level, }{Header level associated with the section header}
}
\description{
Quarto tabset with section header
}
