% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{section_header}
\alias{section_header}
\title{Insert section header}
\usage{
section_header(title, level)
}
\arguments{
\item{title}{Header title}

\item{level}{Numeric specfying the header level}
}
\description{
Places a section header in a quarto document
}
