% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_div}
\alias{quarto_div}
\alias{quarto_span}
\title{Quarto divs and spans}
\usage{
quarto_div(.content, .class = NULL, .sep = "")

quarto_span(.content, .class = NULL, .sep = "")
}
\arguments{
\item{.content, }{Content to be contained in the div/span}

\item{.class, }{CSS class (or vector of classes) for the div/span}

\item{.sep, }{Separator used when merging content}
}
\description{
Quarto divs and spans
}
