% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_group}
\alias{quarto_group}
\alias{quarto_markdown}
\title{Quarto groups}
\usage{
quarto_group(...)

quarto_markdown(..., .sep = "")
}
\arguments{
\item{..., }{Objects to be rendered as a group}

\item{.sep, }{Separator}
}
\description{
Quarto groups
}
