% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_group}
\alias{quarto_group}
\alias{quarto_markdown}
\title{Quarto groups}
\usage{
quarto_group(content, sep = "")

quarto_markdown(content, sep = "")
}
\arguments{
\item{content, }{A list of objects}

\item{sep, }{Separator}
}
\description{
Quarto groups
}
