% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_tabset}
\alias{quarto_tabset}
\title{Quarto tabset}
\usage{
quarto_tabset(.content, .level = 3L)
}
\arguments{
\item{.content, }{Tab title and content as a named list}

\item{.level, }{Header level associated with the tabs}
}
\description{
Quarto tabset
}
