% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_tabset}
\alias{quarto_tabset}
\title{Quarto tabset with section header}
\usage{
quarto_tabset(content, level, title = NULL, names = NULL)
}
\arguments{
\item{content, }{Tab content as a list}

\item{level, }{Header level}

\item{title, }{Section title (optional)}

\item{names, }{Names for tabs (defaults to names(content))}
}
\description{
Quarto tabset with section header
}
\examples{
quarto_tabset(list(tab1 = 1:10, tab2 = "hello"), level = 3L) 

}
