% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{quarto_tabset}
\alias{quarto_tabset}
\title{Quarto tabset with section header}
\usage{
quarto_tabset(.content, .level, .title = NULL)
}
\arguments{
\item{.content, }{Tab titles and content as named list}

\item{.level, }{Header level}

\item{.title, }{Section title (optional)}
}
\description{
Quarto tabset with section header
}
