% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabs}
\alias{tabs}
\title{Insert a tabset}
\usage{
tabs(..., .level = 3L)
}
\arguments{
\item{..., }{Contents of each tab as name-value pairs}

\item{.level, }{Depth level for the implied section headers}
}
\description{
Opens and closes a tabs section in a quarto document
}
